crosswalk/classifier/requirements.txt
# GenAI Security Crosswalk — Classifier Pipeline
# Pin versions for reproducibility

sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
